+++ tests/__init__.py
@@ -7161,12 +7161,10 @@ def f7(param, param2, param3) -> None:
     {E[105].fstring(T)}
     {E[201].fstring(T).format(directive=UNKNOWN)}
 {PATH}:52 in f6
-    {E[101].fstring(T)}
     {E[102].fstring(T)}
     {E[106].fstring(T)}
     {E[201].fstring(T).format(directive=UNKNOWN)}
 {PATH}:62 in f7
-    {E[101].fstring(T)}
     {E[102].fstring(T)}
     {E[106].fstring(T)}
     {E[107].fstring(T)}
@@ -7270,12 +7268,10 @@ def f7(param, param2, param3) -> None:
     {E[105].fstring(T)}
     {E[201].fstring(T).format(directive=UNKNOWN)}
 {PATH}:52 in f6
-    {E[101].fstring(T)}
     {E[102].fstring(T)}
     {E[106].fstring(T)}
     {E[201].fstring(T).format(directive=UNKNOWN)}
 {PATH}:62 in f7
-    {E[101].fstring(T)}
     {E[102].fstring(T)}
     {E[106].fstring(T)}
     {E[107].fstring(T)}