+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/docsig/compare/v0.53.1...HEAD)
+[Unreleased](https://github.com/jshwi/docsig/compare/v0.53.2...HEAD)
+------------------------------------------------------------------------
+
+[0.53.2](https://github.com/jshwi/docsig/releases/tag/v0.53.2) - 2024-05-09
 ------------------------------------------------------------------------
 ### Fixed
 - raise E102 before E107 considered